demo/requirements.txt
flask==3.1.0
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
google-cloud-trace==1.16.1
google-cloud-profiler==4.1.0